feat: polish rankings page with i18n and public access
- Add full i18n support across en/de/es/ja locales
- Refactor to server/client component pattern with breadcrumb JSON-LD
- Polish UI with stat cards, change badges, hover states, and rounded containers
- Make /rankings a public route, no longer behind auth
